release: Fix connection of user (#303)

+++ src/app/commons/services/authentication.service.ts
@@ -10,6 +10,9 @@ import { CheckGrade } from '../utils/check-grade.util';
 import { DateCustom } from '../utils/date.custom';
 import { AuthenticationClientService } from './authentication-client.service';
 
+/**
+ * Service for authentication
+ */
 @Injectable({
     providedIn: 'root',
 })
@@ -27,31 +30,39 @@ export class AuthenticationService {
     ) {}
 
     /**
-     * Login the user
+     * Logs in the user
      */
     public login(): void {
-        this.authenticationClient.login().subscribe((token: Connection): void => {
-            const user: Member | undefined = this.membersApi.isClanMembers(
-                Number(token.account_id)
-            );
+        let token: Connection;
+        let user: Member;
 
-            this.updateStore(user, token?.access_token);
+        this.authenticationClient.login().subscribe({
+            next: (connection: Connection): void => {
+                console.log('Wot api authentication success');
+                token = connection;
+                user = this.membersApi.isClanMembers(Number(connection.account_id));
+            },
+            error: err => {
+                console.error(err);
+            },
+            complete: (): void => {
+                this.updateStore(user, token?.access_token);
 
-            if (user) {
-                this.cookieService.set(CookieNameEnum.TOKEN_WARGAMING, JSON.stringify(token));
-                this.cookieService.set(
-                    CookieNameEnum.TOKEN_USER,
-                    JSON.stringify(user),
-                    DateCustom.getMidnightDate()
-                );
-            }
-            this.router.navigate(['/']).then((): void => {});
+                if (user) {
+                    console.log('User found and authenticated');
+                    this.cookieService.set(CookieNameEnum.TOKEN_WARGAMING, JSON.stringify(token), DateCustom.getToWeeks());
+                    console.log(`Cookie ${CookieNameEnum.TOKEN_WARGAMING} created`);
+                    this.cookieService.set(CookieNameEnum.TOKEN_USER, JSON.stringify(user), DateCustom.getToWeeks());
+                    console.log(`Cookie ${CookieNameEnum.TOKEN_USER} created`);
+                }
+                this.router.navigate(['/']).then((): void => {});
+            },
         });
     }
 
     /**
-     * Check if the user is connected of not
-     * If we found the cookie, we auto connect the user
+     * Checks if the user is logged in or not
+     * If the user's data is found in the cookie, the user is automatically logged in
      */
     public isLoggedIn(): boolean {
         const token: string = this.cookieService.get(CookieNameEnum.TOKEN_WARGAMING);
@@ -63,27 +74,32 @@ export class AuthenticationService {
             if (token) {
                 this.authenticationClient.logout(JSON.parse(token)?.access_token).subscribe({
                     complete: (): void => {
-                        this.cookieService.delete(CookieNameEnum.TOKEN_WARGAMING);
+                        if (this.cookieService.check(CookieNameEnum.TOKEN_WARGAMING)) {
+                            this.cookieService.delete(CookieNameEnum.TOKEN_WARGAMING);
+                        }
+                        if (this.cookieService.check(CookieNameEnum.TOKEN_USER)) {
+                            this.cookieService.delete(CookieNameEnum.TOKEN_USER);
+                        }
                     },
                 });
             }
             return false;
         }
 
-        this.updateStore(JSON.parse(user) ?? undefined, JSON.parse(token)?.access_token);
+        this.updateStore(JSON.parse(user), JSON.parse(token)?.access_token);
         return true;
     }
 
     /**
-     * Update the member store with the data of the connected user
+     * Updates the member store with the data of the connected user
      * @param user The connected user
      * @param accessToken The Wargaming access token
      * @private
      */
-    private updateStore(user: Member | undefined, accessToken: string | undefined): void {
+    private updateStore(user: Member, accessToken: string | undefined): void {
         this.memberStore.patch({
-            account_id: user?.account_id,
-            isAdmin: CheckGrade.isAdmin(user?.role),
+            user: user,
+            isAdmin: CheckGrade.isAdmin(user.role),
             isVisitor: !user,
             accessToken: accessToken,
         });
